[zircon] Port mexec syscalls to Rust
Ports sys_system_mexec and sys_system_mexec_payload_get to Rust.
Architecture-specific and post-shutdown execution details are kept
in C++ and called via FFI in system_ffi.cc.

Bug: 507423574
Test: mexec-smoke-test, mexec-zbi-hello-world-test
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1715811
Fuchsia-Auto-Submit: James Robinson <jamesr@google.com>
API-Review: Adam Barth <abarth@google.com>
Commit-Queue: James Robinson <jamesr@google.com>

+++ zircon/kernel/lib/syscalls/sampler.rs
@@ -40,7 +40,7 @@ pub fn sys_sampler_create(
 
     check_sampler_supported()?;
 
-    validate_ranged_resource(rsrc, ZX_RSRC_KIND_SYSTEM, ZX_RSRC_SYSTEM_SAMPLING_BASE as usize, 1)?;
+    validate_ranged_resource(rsrc, ZX_RSRC_KIND_SYSTEM, ZX_RSRC_SYSTEM_SAMPLING_BASE, 1)?;
 
     ProcessDispatcher::with_current(|up| {
         up.enforce_basic_policy(ZX_POL_NEW_SAMPLER)?;
